Prefer type which allows semantically sound instance properties
It made little sense for our view controllers to be instantiated with a `text` value of "" (i.e. empty string). In that scenario, it actually *does not* have a string and thus `nil` would make more sense (i.e. the absence of a string, rather than some arbitrary default value).