Provide fake network data when testing
This also allows us to simulate a network that takes a little time by delaying the result by two seconds.

Additionally, this combines the previous implementation (where dummy data for a non-real city was being created in a static function) with implementation to provide (not really) pseudo-random data based on Australian capital cities' numeric identifiers.