Prefix all Weather values with a `nil` value
This is in an attempt to get the Weather view controller to show its activity indicator on every view of its tab. However, there is an issue. Because a new `transitions` observable is created every time a `contentProducer` is set and `transitions` starts by combining with `nil`, not `currentWeather` / `currentModel`, it does not track and does not know about the state of any previous observable's work.

This will be raised as an (relatively straightforward and simple) issue for any collaborators. It is simple because the view controller already captures state surrounding its current model.